1. Cleaned up error codes and string lengths 2. Changed some comments
git-svn-id: https://svn.mcs.anl.gov/repos/cifts/ftb/trunk@385 b452418e-dd38-0410-89fe-f8a55b138f88

+++ src/client_lib/include/ftb_client_lib.h
@@ -4,7 +4,7 @@
 #include "ftb_manager_lib.h"
 #include "ftb_def.h"
 
-int FTBC_Connect(const FTB_client_t *cinfo, int extension, FTB_client_handle_t *client_handle);
+int FTBC_Connect(const FTB_client_t *cinfo, uint8_t extension, FTB_client_handle_t *client_handle);
 
 int FTBC_Declare_publishable_events(FTB_client_handle_t client_handle, const char * schema_file, const FTB_event_info_t  *einfo, int num_events);
